more automated. only builds if remote repo is updated!

+++ BuildScript.sh
@@ -1,17 +1,30 @@
 #!/bin/bash
 
 echo "... Checking git"
-DIR=`pwd`
-UNITYLOC="/Applications/Unity/Unity.app/Contents/MacOS/Unity"
 
-GITFETCH="git fetch"
-GITRESET="eval git reset --hard origin/master"
-eval $GITFETCH
-eval $GITRESET
+LOCAL=$(git rev-parse @)
+REMOTE=$(git rev-parse @{u})
+BASE=$(git merge-base @ @{u})
 
-echo "... Building Unity3d Project"
-BUILD="$UNITYLOC -quit -batchmode -projectPath $DIR -logFile $DIR'/Build.log'  -executeMethod ToBuild.BuildLinux"
-if eval $BUILD
-	then echo "... Build Complete!"
-else echo "... An Error Occurred."
+if [ $LOCAL = $REMOTE ]; then
+    echo "... GIT Up-to-date. New Build not required."
+elif [ $LOCAL = $BASE ]; then
+	
+	echo "... New commit detected! Pulling..."
+	GITFETCH="git fetch"
+	GITRESET="eval git reset --hard origin/master"
+	eval $GITFETCH
+	eval $GITRESET  
+	
+	DIR=`pwd`
+	UNITYLOC="/Applications/Unity/Unity.app/Contents/MacOS/Unity"
+
+
+	echo "... Re-Building Unity3d Project"
+	BUILD="$UNITYLOC -quit -batchmode -projectPath $DIR -logFile $DIR'/Build.log'  -executeMethod ToBuild.BuildLinux"
+	if eval $BUILD
+		then echo "... Build Complete!"
+	else echo "... An Error Occurred."
+	fi
+
 fi
